Documents/python/unlimitedsafari-master/requirements.txt
asgiref==3.2.7
astroid==2.4.0
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
decorator==4.4.2
Django==3.0.5
django-background-tasks==1.2.5
django-compat==1.0.15
django-extensions==2.2.9
djangorestframework==3.11.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
names==0.3.0
parsel==1.5.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
Protego==0.1.16
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.5.0
pyOpenSSL==19.1.0
python-scrapyd-api==2.1.2
pytz==2020.1
queuelib==1.5.0
random-password-generator==2.1.0
requests==2.23.0
Scrapy==2.1.0
scrapyd==1.2.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
splinter==0.13.0
sqlparse==0.3.1
toml==0.10.0
traitlets==4.3.3
Twisted==20.3.0
urllib3==1.25.9
w3lib==1.21.0
wcwidth==0.1.9
wrapt==1.12.1
zope.interface==5.1.0
